Preparetion for creating heuristic function.
For heuristic function we somehow should estimate the remaining distance to the goal. Problem is how to understand weight value. These values do not depend on geometry positions, as i see it. I came to conlusion that we could calcualte weigh for one pixel for each edge and take the smaller value. It will not break the main rule for a heuristic function, it should not produce the value bigger than real.